clean up subscription initialization tests
