+++ momento-sdk/build.gradle.kts
@@ -110,8 +110,3 @@ registerIntegrationTestTask(
     "test-retries",
     listOf("momento.sdk.retry.*")
 )
-
-registerIntegrationTestTask(
-    "test-topics-subscription-initialization",
-    listOf("momento.sdk.subscriptionInitialization.*")
-)
